[ios] don't limit possible orientations for <Modal>
It's unclear why the ViewController for <Modal> restricts the screen orientations on iPhone, but it's also pretty clear from its comment that whoever made that choice realized that there needs to be some way for consumers to set their own preferences here. I propose that such a preference should not be baked into the ViewController but left up to the consumer using a custom mechanism or something like react-native-orientation.