Update Android Performance for opening trace.html
Summary:
Right now systrace generates an html file that is using the Object.observe function which is now deprecated in most up to date browsers.

Since we don't want to depend on the systrace being up to date, we should advise opening the trace through the Tracing tool, since that is the only way to get it working until systrace removes its dependency on Object.observe.
Closes https://github.com/facebook/react-native/pull/7880

Differential Revision: D3516305

Pulled By: astreet

fbshipit-source-id: ab9ae46ba74b7a34b6cfb8c37ca1ec2e6b41e353