Android: Enable views to be nested within <Text>
Potential breaking change: The signature of ReactShadowNode's onBeforeLayout
method was changed
  - Before: public void onBeforeLayout()
  - After:  public void onBeforeLayout(NativeViewHierarchyOptimizer nativeViewHierarchyOptimizer)

Depends on this css-layout PR:
https://github.com/facebook/css-layout/pull/202

Implements same feature as this iOS PR:
https://github.com/facebook/react-native/pull/7304

Previously, only Text and Image could be nested within Text. Now, any view can
be nested within Text. One restriction of this feature is that developers must
give inline views a width and a height via the style prop.

Previously, inline Images were supported via
FrescoBasedReactTextInlineImageSpan. To get support for nesting views within
Text, we create one special kind of span per inline view. This span is called
TextInlineViewPlaceholderSpan. It is the same size as the inline view. Its job
is just to occupy space -- it doesn't render any visual. After the text is
rendered, we query the Android Layout object associated with the TextView to
find out where it has positioned each TextInlineViewPlaceholderSpan. We then
position the views to be at those locations.

One tricky aspect of the implementation is that the Text component needs to be
able to render native children (the inline views) but the Android TextView
cannot have children. This is solved by having the native parent of the
ReactTextView also host the inline views. Implementation-wise, this was
accomplished by extending the NativeViewHierarchyOptimizer to handle this case.
The optimizer now handles these cases:
  - Node is not in the native tree. An ancestor must host its children.
  - Node is in the native tree and it can host its own children.
  - (new) Node is in the native tree but it cannot host its own children. An
    ancestor must host both this node and its children.

Limitation: Clipping
====
If Text's height/width is small such that an inline view doesn't completely fit,
the inline view may still be fully visible due to hoisting (the inline view
isn't actually parented to the Text which has the limited size. It is parented
to an ancestor which may have a different clipping rectangle.). Prior to this
change, layout-only views had a similar limitation.