Don't clear ReactNativeHost if other inst exists
Do not call `getReactNativeHost().clear();` in `ReactActivity.onDestroy` if another created `ReactActivity` instance exists. This avoids the scenario where a dying `ReactActivity` in the background clears and destroys the shared `ReactInstanceManager` instance being used by another `ReactActivity` in the foreground. This fixed a scenario which can result in lifecycle corruption.