Fix maxLength issue of TextInput on MarkedRange and AutoCompletion
MaxLength doesn't work correctly with MarkedRange and AutoCompletion.
In some cases,
textField:shouldChangeCharactersInRange:replacementString: will not be
called when using AutoCompletion. And, we should allow user to type
when the confirmed string length is not exceed the limitation of
maxLength. So we need to check the length of text when we receive
UIControlEventEditingChanged event.