Prevent race condition on immediate transition
NavigationTransitioner prepares for transition within componentWillReceiveProps, using previously-saved state to determine how to properly handle new props. If a transition is to take place, the code saves new info in state, executes the transition, and cleans up scenes within _onTransitionEnd.

If the transition is a jump-to transition, or otherwise takes very little time, then it is possible for the setState call within _onTransitionEnd to use state which hasn't yet been set by the code within componentWillReceiveProps, resulting in a failed transition.

This fix ensures that the initial setState call is completed before executing the transition.