Eradicate the 'unique "key" prop' warnings
For the guides and Components/APIs, we would continuously get:

  Warning: Each child in an array or iterator should have a unique "key" prop.

when rendering them.

This add `key` attributes to our DOM elements that did not have them in order
to remove this warning. It uses a random string generator to create unique keys.

> Should we establish the key a different or better way?

I went full hammer on this and put `key` attributes on everything. May have
been overkill, but the warnings are gone.

> Should I be more selective on what we put the `key` attribute on?

Also fixed a:

`Warning: Unknown DOM property frameborder. Did you mean frameBorder?`

issue as well.

Test Plan:

- No more warnings
- Guides still render the same
- APIs still render the same (when applied with fix from
https://github.com/facebook/react-native/pull/8746)