Add push action support to PushNotificationIOS
This fixes some issues I've had with the PushNotificationIOS library.
The biggest change is adding support for push actions.

Currently it's impossible to schedule a silent local notification, since empty
notification sounds are replaced with the default sound. This commit deprecates
this use of the API in favor of passing an explicit sound name and adding a
getDefaultSoundName() method.

Test plan: This code is already running in production; just cleaned up
to merge upstream.