Use sequential, deterministic keys instead of random
I had actually tried to reduce the number of `key` attributes on the
elements, but there was a lot of cases where I thought I had fixed
all the warning, but then clicked on another doc and the warning would
come back because a new element that did not have the `key` prop ended
up being in that doc.

If we really think having `key` props on all the elements are bad, is
there an easy way to test to see which ones are really necessary?