Add capability for MessageQueue to emit "SPY" events in a way that can be extensible.
After this change, we can wire a "spy" into a queue
that will expose the events in different and interesting
ways. This removes the hardcoded `SPY_MODE` flag and
instead uses a function that can be injected from outside.

This change will enable a tool like Snoopy to exist,
without its internal monkeypatching of MessageQueue:

https://github.com/jondot/rn-snoopy

This was a low-hanging-fruit waiting to be picked. And
now, we have deeper and more elaborate introspection
at the React Native bridge traffic, empowering optimization
and debugging scenarios alike.