Stop Reload Android Webview On Same URL Before
### Motivation
Because it is react, the url could change on redirection or some other ways.
the iOS version's WebView has control that on [here](https://github.com/facebook/react-native/blob/master/React/Views/RCTWebView.m#L106).
But the Android's one is not.

### What this commit do
Check the url is same with privous url. If it is true, cancel loading.
This logic is same with iOS's.

### What should we know
the ```method``` hasn't compared.