Fix CLI help when examples property is present.
When an `examples` property is present, the command is assumed to not take any additional command line arguments (i.e., the usage is `[options]` only), and the return value of `.usage()` is an array of examples.

This provides a proper check for the result of `cmd.usage()` to be a string (the normal case when examples are not present), in which case the set of examples are not processed.

Prior to this change, running `react-native link --help` would result in an exception, because the `.usage()` output for `link` is `[options] [packageName]` (note, !== `[options]`).

In general, overriding the behavior of usage() may be limiting, as this change will not support adding examples to `react-native link --help` while still allowing the first line of the --help message to include the original usage (`[options] [packageName]`).