Fix memory leak in MessageQueue
The MessageQueue has a _debugInfo object where it stores
debug information associated with each callback. The size
of this structure is currently unbounded.

It looks like the code attempted to restrict _debugInfo to
a fixed number of entries but due to a logic bug, it leaked
around 30 entries for every 1 entry it cleaned up.

This change limits the _debugInfo object to around 30 entries.