Further RNPM integration
This commit removes `rnpm` folder that we left during initial merge to
keep the diff cleaner. There's now:

1) core part - it will be further integrated with the `utils` folder
(esp. the config part that I am working on concurrently)
2) link
3) install

The 2 and 3 reproduce the same structure as in any other command.

From most notable differences, the `src` folder is now gone. The new
structure should make it easier for people to work with the stuff and
also move us closer to 100% rnpm integration.