Use mergeFast in a hotspot
In profiling our app, we found that the usage
of `merge` in `Text.js` was showing up as a
hotspot. We've replaced this usage of `merge`
with `mergeFast`.

We've also updated `mergeFast` to use `merge`
in dev mode so that developers can benefit
from the extra validation it does on its
arguments.