Catch Throwable instead of Exception from loadLibrary
We're seeing cases where RN initialization fails due to:
`UnsatisfiedLinkError: could find DSO to load: libreactnativejni.so`
We're using `NativeModuleCallExceptionHandler` to catch initialization
exceptions and prevent the app from crashing. However, since
`UnsatisfiedLinkerError` is an `Error` instead of `Exception`, it is not
caught and thus the app crashes anyways.
This change fixes that by catching `Throwable` instead of `Exception`
from `XReactInstanceManagerImpl.doInBackground()`. I also had to change
`Result.of()` to take a `Throwable` as well, thus a few other places had
to be updated as well.