stopPropagation on navigationComplete event. Fixes #1241
With nested NavigatorIOS components:

```
<NavigatorIOS initialRoute={{
  component: ComponentWithNavigatorIOSInSubtree
  title: 'xyz'
}}>
```

Navigating (via push etc.) in ComponentWithNavigatorIOSInSubtree's `navigator` caused an invariant error, making apps with this structure unusable. The reason was that the `navigationComplete` nativeEvent was being propagated to the outer Navigator due to React's automatic event bubbling. The event bubbling was automatic and almost certain unintentional, so preventing it should not cause unintended side-effects.

See demo repo at #1241 for more a more detailed example.