[Packager] Add "assetPlugin" option to allow arbitrary asset processing
Adds a new URL option to the packager server called "assetPlugin". This can be a name of a Node module or multiple Node modules (`assetPlugin=module1&assetPlugin=module2`). Each plugin is loaded using `require()` and is expected to export a function. Each plugin function is invoked with an asset as the argument. The plugins may be async functions; the packager will properly wait for them to settle and will chain them.

A plugin may be used to add extra metadata to an asset. For example it may add an array of hashes for all of the files belonging to an asset, or it may add the duration of a sound clip asset.

Test Plan: Added unit tests for Server and Bundler that test URL option parsing and invocation of the plugins. Tested that plugins run in order and that async plugins are supported.