[ResolutionRequest] Only push 'node_modules' directories to the `searchQueue`
This fixes a specific edge case (and avoids the need to use `searchQueue.map`).

The edge case:

If an import path contains `path.sep`, using `path.dirname` with `searchQueue.map` won't return the expected value (eg: '/path/to/node_modules').

To fix that, we only push the 'node_modules' directories to the `searchQueue`. Once we iterate the `searchQueue`, we can append the `realModuleName` to each `searchPath`.