Simplify react-native overrides when building from source
Summary:
You may want to double-check that this works perfectly (I am trying to move away from RN-from-source, since my changes were upstreamed), but it seems to build without errors for me. And this has the advantage of working seamlessly with "rnpm", which isn't smart enough to add these excludes when linking modules.
Closes https://github.com/facebook/react-native/pull/9954

Differential Revision: D4074059

fbshipit-source-id: e152945ea66605698c18c1d0d078fbf312f1658e