Summary:
Commit https://github.com/facebook/react-native/commit/3d3b067f6fc831b6b23726087fe39cf39ef86f00 implemented zIndex support for Android. However after zIndex resorting of the views, the JS doesn't know about the new order. When dynamically removing the view from an array of views, JS sends the command to remove the view at certain index, which is not the same in Java since the order of items changed.

Test Plan:
Related issue and the simple empty react-native 0.37 app where it was tested:
https://github.com/facebook/react-native/issues/8968

Initially the order of operations in `manageChildren` in `NativeViewHierarchyManager.java` was:
1) Loop by `indicesToRemove`, call `viewManager.removeViewAt(viewToManage, indexToRemove)`. However if the view is delegated and it's ID in the array of `tagsToDelete`, do nothing and delegate this job of removing item to the step 3).
2) Loop by `viewsToAdd`, add the views.
3) Loop by `tagsToDelete`, perform animations on the views that are using animations, then call `viewManager.removeView(viewToManage, viewToDestroy)` and `dropView(viewToDestroy)`. But if the view is not animated, only `dropView(viewToDestroy)` is called because we assume that correct view was previously removed in step 1)

New order of operations:
1) Loop by `tagsToDelete`, perform animations if necessary, then call `viewManager.removeView(viewToManage, viewToDestroy)` and `dropView(viewToDestroy)`. If no animations, just call `viewManager.removeView(viewToManage, viewToDestroy)` and `dropView(viewToDestroy)` right away.
2) Loop by `viewsToAdd`, add the views.

The intent is to perform clean-up by only looping by `tagsToDelete` first, instead of splitting the clean-up into separate steps of removing by index, then dropping by tagToDelete. And then when all the views are properly removed, we add the new views. Since `indicesToRemove` it's not reliable anymore and we don't use it for `removeViewAt`, we can get rid of the whole 1) step and replace it by 3) step.
Also by having the old operations order we having the case that the *wrong* view is removed in step 1), but then, *correct* view is dropped in step 3). In result we have inconsistency when removing one view but dropping another. It's described in more details in mentioned above related issue comments.

Because the diff doesn't look very clear because of replacing the blocks of code, here is what's actually changed:
1. Removed the loop by `indicesToRemove` as it's not reliable anymore and causing wrong view to be removed.
2. Moved loop by `tagsToDelete` before adding new views
3. Added a call for `viewManager.removeView(viewToManage, viewToDestroy)` in `else` statement.