Android: Introduce viewLayerTypeAndroid and deprecate renderToHardwareAndroid
React Native for Android supports the renderToHardwareAndroid prop which enables a developer to set a view's layer type to LAYER_TYPE_HARDWARE or LAYER_TYPE_NONE. Natively, Android supports 3 values for a view's layer type but only 2 of them are exposed to JavaScript.

This change introduces a prop called viewLayerTypeAndroid which is an enum which exposes all 3 possible view layer types to JavaScript:
  - LAYER_TYPE_NONE
  - LAYER_TYPE_HARDWARE
  - LAYER_TYPE_SOFTWARE

This change also deprecates renderToHardwareAndroid in favor of viewLayerTypeAndroid because viewLayerTypeAndroid is a superset of renderToHardwareAndroid.