iOS: Ensure correct width/height after rotation
On some devices, we are seeing that the width and height reported by `didUpdateDimensions` are swapped.

Here's the problem. Inside of the `UIApplicationWillChangeStatusBarOrientationNotification` handler, on some devices we are seeing that the dimensions reported by `[[UIScreen mainScreen] bounds]` have already been updated to reflect the new orientation while on others they have not been.

To fix this, `RCTExportedDimensions` now manually ensures the dimensions it returns match the desired orientation.