Remove deprecated push notification method info
The method didReceiveRemoteNotification without a fetch completion
handler is deprecated by Apple and they discourage using it:
https://developer.apple.com/reference/uikit/uiapplicationdelegate/1623117-application

Also, there is a practical consideration for it. The version without
the handler *only works when application is in the foreground* which
is not what most of the people expect from their push notifications.

The version with the handler works both in the foreground and
background.