Custom interpolator prop for NavigationCardStack
This makes it possible override the built in card stack interpolator without having to maintain your own copy of NavigationCardStack. We were not crazy about the slight zooming or the opacity change that the built-in one was doing for the previous card, so this allows us to just use the translate.

This leaves the current behavior and API completely intact, but adds a new prop to NavigationCardStack for passing in your own interpolator. People that want to do this will want to look at the built-in one to get an idea of how an interpolator works.