Mention in the docs how to install JS dependencies
I don't think this is the right place, but either I have completely missed it, or there is no mention of installing JS libraries anywhere in docs. It works perfectly out of the box, but I think it deserves a concrete mention (along the lines of what I added). Perhaps it should be a separate page under "Guides"? I don't have a good solution, just want to raise this.