Add integration tests
The reload and force-cache tests don't actually verify that the complete functionality.

reload: Should have the server set a long cache header, then swap the image on next load
with the test comparing the old image to the new image and making sure they are different.

force-cache: Should do the above but set a no-cache header. The test should compare the first
image with the new one and make sure they are the same.