Upgrade glob to v7.1.1
There are three breaking changes:

- 6.0
  - Remove comment and negation pattern support
  - Ignore patterns are always in dot:true mode
- 7.0
  - Raise error if options.cwd is specified, and not a directory

I ran `npm run test`, and all tests passed locally. I also quickly
grepped the codebase, to see if it used any of those breaking changes.

There are three places that specify `cwd`:

1. `local-cli/core/android/findPackageClassName.js:22`
  - `const files = glob.sync('**/*.java', { cwd: folder });`

2. `local-cli/core/windows/findNamespace.js:22`
  - `const files = glob.sync('**/*.cs', { cwd: folder });`

3. `local-cli/core/windows/findPackageClassName.js:22`
  - `const files = glob.sync('**/*.cs', { cwd: folder });`

Nothing appeared to use negation or comment patterns in a glob,
from my limited explorations, and there did not seem to be anything 
that set the `dot` argument.

## Motivation
$ npm ls glob
├─┬ eslint@3.13.1
│ └── glob@7.1.1
├─┬ jest@18.1.0
│ └─┬ jest-cli@18.1.0
│   └─┬ istanbul-api@1.1.0
│     ├─┬ fileset@2.0.3
│     │ └── glob@7.1.1
│     ├─┬ istanbul-lib-report@1.0.0-alpha.3
│     │ └─┬ rimraf@2.5.4
│     │   └── glob@7.1.1
│     └─┬ istanbul-lib-source-maps@1.1.0
│       └─┬ rimraf@2.5.4
│         └── glob@7.1.1
└─┬ react-native@0.39.2
  ├── glob@5.0.15
  └─┬ rimraf@2.5.4
    └── glob@7.1.1

Of the three dependencies I have that rely on `glob`, react-native is the only one not on 7.x. If react-native upgrades, then npm can deduplicate the glob dependency.